Revery

Framework for fast, native code, cross-platform GUI applications.

Overview

Revery is a framework for building cross-platform GUI applications. Revery provides a React-like, functional approach for modeling UI, as well as scaffolding for managing the application lifecycle.

Revery started as the foundation of Onivim 2, but was factored out into a general toolkit for ReasonML user interfaces.

Quickstart

There are two ways to get started:

Try out the revery Playground

Clone and run revery-quick-start

git clone https://github.com/revery-ui/revery-quick-start

cd revery-quick-start

esy install

esy build

esy run

Component Model

Basic Components

The basic component building block is simply a pure function, that returns an element.

Example:

let squareBox = () => <Container width=10 height=10 color=Colors.red />;

Properties can be specified as named arguments.

Example:

let wideBox = (~width: int, ()) => <Container width height=10 color=Colors.red />;

You can accept children through through the ~children property, for example:

let boxWithChildren = (~children, ()) => <Container width=10 height=10 color=Colors.red>children</Container>;

Components with Hooks

For components that manage state, or have side effects, you'll want to encapsulate those using hooks!

A component that uses hooks can be created using let%component, for example:

let%component componentWithHooks = (~children, ()) => {
  let%hook (state, setState) = Hooks.state(0);

  <Button text=string_of_int(state) onClick={(_) => setState(state => state + 1)}/>;
};

Hooks components still return a function with named arguments representing the properties, but behind the scenes a hooks component is a function that takes a hooks object, and then returns a tuple of (hooks, element). This complexity is hidden away for you using syntax sugar in a manner similar to how async/await hides promises in JavaScript for example.

In effect, Revery components should always be pure functions - given a set of props, and a hooks, the output should always be the same.