Module Revery.UI

include Revery_UI

module Animation = Revery_UI.Animation

module Spring = Revery_UI.Spring

module Easing = Revery_UI.Easing

module ImageResizeMode = Revery_UI.ImageResizeMode

module Layout = Revery_UI.Layout

module LayoutTypes = Layout.LayoutTypes

module Style = Revery_UI.Style

module Transform = Revery_UI.Transform

module Selector = Revery_UI.Selector

class node : unit -> Revery_UI__.Node.node

class viewNode : unit -> Revery_UI__.ViewNode.viewNode

class textNode : string -> Revery_UI__.TextNode.textNode

class imageNode : string -> Revery_UI__.ImageNode.imageNode

class canvasNode : unit -> Revery_UI__.CanvasNode.canvasNode

module NodeDrawContext = Revery_UI.NodeDrawContext

module Keyboard = Revery_UI.Keyboard

module Mouse = Revery_UI.Mouse

module NodeEvents = Revery_UI.NodeEvents

module UiEvents = Revery_UI.UiEvents

module Container = Revery_UI.Container

module React = Revery_UI.React

module Focus = Revery_UI.Focus

module Dimensions = Revery_UI.Dimensions

module Offset = Revery_UI.Offset

type element = node React.element

include Revery_UI__.Ui

type renderFunction = Revery_UI.React.reveryNode Revery_UI__.React.element -> unit

val start : Revery_Core.Window.t -> Revery_UI.React.reveryNode Revery_UI__.React.element -> renderFunction

start is the entry point for creating a UI driven by the brisk-reconciler React infrastructure.

This does the following:

Hook up a render loop

Take a React.element and render it via GPU

Hook up GLFW events and convert them to Revery UI events

start returns a function that can be used to update the UI.

An example usage would be: let update = start(window, <View />); update(<Button />);

val getActiveWindow : unit -> Revery_Core.Window.t option

include Revery_UI_Primitives

module BoxShadow = Revery_UI_Primitives.BoxShadow

module PrimitiveNodeFactory = Revery_UI_Primitives.PrimitiveNodeFactory

module AllowPointer = Revery_UI_Primitives.AllowPointer

module IgnorePointer = Revery_UI_Primitives.IgnorePointer

module Image = Revery_UI_Primitives.Image

module Canvas = Revery_UI_Primitives.Canvas

module Opacity = Revery_UI_Primitives.Opacity

module Padding = Revery_UI_Primitives.Padding

module Text = Revery_UI_Primitives.Text

module View = Revery_UI_Primitives.View

module Components = Revery_UI_Components

module Hooks = Revery_UI_Hooks