Module Revery_Core.Environment

val isNative : bool

Environment provides information about the run-time environment

true if native / bytecode build, false if JSOO build

val webGL : bool

true if JSOO build, false otherwise

val sleep : Time.t -> unit

sleep(t) sleeps the running thread for the specified time period.

Example usage: sleep(Seconds(1.0))

No-op in JSOO builds.

val yield : unit -> unit

val getExecutingDirectory : unit -> string

@deprecated

val executingDirectory : string

executingDirectory gets the directory where the running executable lives.

val getWorkingDirectory : unit -> string

val getAssetPath : string -> string

getAssetPath resolves a path to an absolute path. If the path is not already absolute, it is assumed to be relative to the current binary.

type os =

| Windows

| Mac

| Linux

| Browser

| Unknown

val os : os