Module Revery_Core.WindowCreateOptions

type t = {

resizable : bool;

visible : bool;
If visible is true, the window will be visible immediately upon creation.

If visible is false, the window will be invisible, and will require a call to Window.toString to make it visible. Can be useful to hide transient UI states.

maximized : bool;
If maximized is true, the window will be expanded immediately upon creation.

decorated : bool;
If decorated is true, the window will be framed with the OS-specific adorners. If false, the default chrome behavior on the OS will be disabled and (if you want to maintain that functionality) you must implement it yourself through the mouseBehavior attribute on Views

titlebarStyle : WindowStyles.titlebar;
titlebarStyle sets the appearance of the titlebar. Eventually this will be platform independent, but as of right now, Transparent and Hidden only works on macOS.

opacity : float;
opacity sets the opacity of the window. A value of 1.0 means the window is fully opaque, and a value of 0.0 means the window is completely transparent.

x : [ `Centered | `Absolute of int ];
x is the initial horizontal position of the Window, either `Centered or `Absolute(x), where x is the horizontal pixel coordinate of the left edge of the Window

y : [ `Centered | `Absolute of int ];
y is the initial horizontal position of the Window, either `Centered or `Absolute(x), where y is the vertical pixel coordinate of the top edge of the Window

width : int;
width is the initial horizontal size of the Window, with display scaling applied.

height : int;
height is the initial vertical size of the Window, with display scaling applied.

minimumWidth : int;
minimumWidth is the minimum horizontal size of the Window.

minimumHeight : int;
minimumWidth is the minimum vertical size of the Window.

backgroundColor : Color.t;
backgroundColor specifies the initial Color of the Window

vsync : Vsync.t;
If vsync is true (default), rendering will be synchronized to the monitor refresh rate.

This can decrease tearing, but also limits the frame rate to the monito refresh rate.

icon : string option;

forceScaleFactor : float option;

}

t is the type representing creation options for a window

val create : ?⁠resizable:bool -> ?⁠visible:bool -> ?⁠maximized:bool -> ?⁠decorated:bool -> ?⁠titlebarStyle:WindowStyles.titlebar -> ?⁠opacity:float -> ?⁠x:[ `Absolute of int | `Centered ] -> ?⁠y:[ `Absolute of int | `Centered ] -> ?⁠width:int -> ?⁠height:int -> ?⁠minimumWidth:int -> ?⁠minimumHeight:int -> ?⁠backgroundColor:Color.t -> ?⁠vsync:Vsync.t -> ?⁠icon:string option -> ?⁠forceScaleFactor:float option -> unit -> t

val default : t