Module Revery_Core__.App

type t

t is the type representing the running app instance

There is only ever a single app instance, but an app instance may have multiple windows.

val getWindows : t -> Revery_Core.Window.t list

getWindows(app) returns the list of all open Window.t instances

val quit : ?⁠askNicely:bool -> ?⁠code:int -> t -> unit

quit(~askNicely, ~code, c) causes the App to quit with exit code c

askNicely specifies whether quit should be forced (default false). If true, the canQuit handlers will be run for each window. If false, the canQuit handlers will be ignored.

code specifies the exit code. Defaults to 0.

val isIdle : t -> bool

isIdle(app) returns true if the app is idling, false othwrise

type delegatedFunc = unit -> unit

val runOnMainThread : delegatedFunc -> unit

runOnMainThread(f) schedules the function f to run during the next render frame on the main thread.

If the application is idle, this will cause it to become active, and trigger a re-render.

val flushPendingCallbacks : unit -> unit

flushPendingCallbacks(f) will explicitly run all the callbacks queued up via runOnMainThread.

In general, this should not need to be called by user code, as it happens as part of the application lifecycle.

However, it may be necessary to call this for tests that queue up callbacks via runOnMainThread.

val setCanIdle : (unit -> bool) -> t -> unit

setCanIdle(f, app) registers a callback f. Each frame, f will be called to check if the application can idle. If f() returns false, the app will not idle and will render. This is useful if you have animations active, but in general, you want to allow the app to idle to minimize CPU and battery usage.

type unsubscribe = unit -> unit

val onBeforeQuit : t -> (unit -> unit) -> unsubscribe

onBeforeQuit(app, f) registers a callback [f] that is called prior to quitting

val onIdle : t -> (unit -> unit) -> unsubscribe

onIdle(app, f) registers a callback [f] that is called when the application is idle. This allows you to defer work when the app is not under load - (for example, this is may be a good time to garbage collect). This will be called when multiple frames have passed without requiring a render.

val createWindow : ?⁠createOptions:Revery_Core.WindowCreateOptions.t -> t -> string -> Revery_Core.Window.t

createWindow ~createOptions, app, name creates a new window

val start : (t -> unit) -> unit

start is the entry point for a Revery application. This initiates the Revery application lifecycle, and an app instance (t) is passed to an initialization function.

val initConsole : unit -> unit

initConsole (Windows-only) attaches or allocates a console, to show logging output. No-op on other platforms.