Module Revery_Core__.Time

type t

val zero : t

val seconds : int -> t

val milliseconds : int -> t

val ms : int -> t

val ofFloatSeconds : float -> t

ofFloatSeconds(seconds) converts a float value in seconds to a t representing the time.

val toFloatSeconds : t -> float

toFloatSeconds(time) converts the time t to a float value representing the time in seconds.

val (+) : t -> t -> t

val (-) : t -> t -> t

val (*) : t -> t -> t

val (*.) : t -> float -> t

val (/) : t -> t -> t

val (/.) : t -> float -> t

val max : t -> t -> t

val min : t -> t -> t

val (mod) : t -> t -> t

val toString : t -> string

val now : unit -> t

now() returns the current system time t