Module Revery_Core__Window

type t

type size = {

width : int;

height : int;

}

type unsubscribe = unit -> unit

val onBeforeRender : t -> (unit -> unit) -> unsubscribe

val onAfterRender : t -> (unit -> unit) -> unsubscribe

val onBeforeSwap : t -> (unit -> unit) -> unsubscribe

val onAfterSwap : t -> (unit -> unit) -> unsubscribe

val onFocusGained : t -> (unit -> unit) -> unsubscribe

val onFocusLost : t -> (unit -> unit) -> unsubscribe

val onMaximized : t -> (unit -> unit) -> unsubscribe

val onFullscreen : t -> (unit -> unit) -> unsubscribe

val onMinimized : t -> (unit -> unit) -> unsubscribe

val onRestored : t -> (unit -> unit) -> unsubscribe

val onCompositionStart : t -> (unit -> unit) -> unsubscribe

val onCompositionEdit : t -> (Revery_Core.Events.textEditEvent -> unit) -> unsubscribe

val onCompositionEnd : t -> (unit -> unit) -> unsubscribe

val onExposed : t -> (unit -> unit) -> unsubscribe

val onKeyDown : t -> (Revery_Core.Key.KeyEvent.t -> unit) -> unsubscribe

val onKeyUp : t -> (Revery_Core.Key.KeyEvent.t -> unit) -> unsubscribe

val onMouseEnter : t -> (unit -> unit) -> unsubscribe

val onMouseLeave : t -> (unit -> unit) -> unsubscribe

val onMouseMove : t -> (Revery_Core.Events.mouseMoveEvent -> unit) -> unsubscribe

val onMouseDown : t -> (Revery_Core.Events.mouseButtonEvent -> unit) -> unsubscribe

val onMouseUp : t -> (Revery_Core.Events.mouseButtonEvent -> unit) -> unsubscribe

val onMouseWheel : t -> (Revery_Core.Events.mouseWheelEvent -> unit) -> unsubscribe

val onSizeChanged : t -> (size -> unit) -> unsubscribe

val onMoved : t -> ((int * int) -> unit) -> unsubscribe

val onTextInputCommit : t -> (Revery_Core.Events.textInputEvent -> unit) -> unsubscribe

val onFileDropped : t -> (Revery_Core.Events.fileDropEvent -> unit) -> unsubscribe

val canQuit : t -> bool

val getDevicePixelRatio : t -> float

val getSize : t -> size

getSize(window) returns a size describing the window dimensions, accounting for display scaling.

val getFramebufferSize : t -> size

val setSize : width:int -> height:int -> t -> unit

setSize(~width, ~height, window) sets the window size, taking display scaling into account.

For example, for a Windows display with 200% scaling, setSize(~width=400, ~height=300, window), will actually set the window to a height of 800 pixels wide by 600 pixels high. This is usually the behavior you want. To directly set the size, without considering display scaling, use setUnscaledSize

val setMinimumSize : width:int -> height:int -> t -> unit

setMinimumSize(~width, ~height, window) sets the minimum window size.

val getPosition : t -> int * int

val getScaleAndZoom : t -> float

val getSdlWindow : t -> Sdl2.Window.t

val getZoom : t -> float

val getUniqueId : t -> int

val getBackgroundColor : t -> Revery_Core.Color.t

val isDirty : t -> bool

val isDecorated : t -> bool

val setInputRect : t -> int -> int -> int -> int -> unit

val startTextInput : t -> unit

val stopTextInput : t -> unit

val isTextInputActive : t -> bool

val center : t -> unit

val hide : t -> unit

val show : t -> unit

val maximize : t -> unit

val isMaximized : t -> bool

val isFullscreen : t -> bool

val raise : t -> unit

val restore : t -> unit

val minimize : t -> unit

val setOpacity : t -> float -> unit

val setBackgroundColor : t -> Revery_Core.Color.t -> unit

val setPosition : t -> int -> int -> unit

val setTitle : t -> string -> unit

val setZoom : t -> float -> unit

val setVsync : t -> Revery_Core.Vsync.t -> unit

val render : t -> unit

val handleEvent : Sdl2.Event.t -> t -> unit

val create : string -> Revery_Core.WindowCreateOptions.t -> t

create(name, options) creates a new Revery application window.

See WindowCreateOptions for a list of available options.

val takeScreenshot : t -> string -> unit

val setCanQuitCallback : t -> (unit -> bool) -> unit

val setRenderCallback : t -> (unit -> unit) -> unit

val setShouldRenderCallback : t -> (unit -> bool) -> unit

val getFPS : t -> int

val showFPSCounter : t -> unit

val hideFPSCounter : t -> unit

val shouldShowFPSCounter : t -> bool