Module Revery_IO__File

val write : path:string -> string -> (unit, string) Stdlib.result Lwt.t

* write * * Takes a path and writes data to that file. * * Examples: * let result: Lwt.t(result(unit, string)) = File.write(~path="example.txt", "Hello World!");

val delete : string -> (unit, string) Stdlib.result Lwt.t

* delete * * Takes a path and tries to delete that file. * * Examples: * let result: Lwt.t(result(unit, string)) = File.delete("example.txt");