Module Revery_UI

module Animation : sig ... end

module Spring : sig ... end

module Easing : sig ... end

module ImageResizeMode : sig ... end

module Layout : sig ... end

module LayoutTypes = Layout.LayoutTypes

module Style : sig ... end

module Transform : sig ... end

module Selector : sig ... end

class node : unit -> Revery_UI__.Node.node

class viewNode : unit -> Revery_UI__.ViewNode.viewNode

class textNode : string -> Revery_UI__.TextNode.textNode

class imageNode : string -> Revery_UI__.ImageNode.imageNode

class canvasNode : unit -> Revery_UI__.CanvasNode.canvasNode

module NodeDrawContext : sig ... end

module Keyboard : sig ... end

module Mouse : sig ... end

module NodeEvents : sig ... end

module UiEvents : sig ... end

module Container : sig ... end

module React : sig ... end

module Focus : sig ... end

module Dimensions : sig ... end

module Offset : sig ... end

type element = node React.element

include Revery_UI__.Ui

type renderFunction = React.reveryNode Revery_UI__.React.element -> unit

val start : Revery_Core.Window.t -> React.reveryNode Revery_UI__.React.element -> renderFunction

start is the entry point for creating a UI driven by the brisk-reconciler React infrastructure.

This does the following:

Hook up a render loop

Take a React.element and render it via GPU

Hook up GLFW events and convert them to Revery UI events

start returns a function that can be used to update the UI.

An example usage would be: let update = start(window, <View />); update(<Button />);

val getActiveWindow : unit -> Revery_Core.Window.t option

type mouseBehavior = Sdl2.Window.hitTestResult =

| Normal

| Draggable

| ResizeTopLeft

| ResizeTop

| ResizeTopRight

| ResizeRight

| ResizeBottomRight

| ResizeBottom

| ResizeBottomLeft

| ResizeLeft

mouseBehavior is an alias for SDL2's hit-test results. Each option is fairly self explanatory, but please note that these only work if you set `decorated` to false on window creation. This is unfortunately due to the fact that enabling hit tests causes the default chrome behavior on Windows to no longer work.