Module Revery_UI.Selector

type 'a selector =

| Color : Revery_Core.Color.t selector

| FontFamily : string selector

| FontSize : float selector

| Width : int selector

| FlexGrow : int selector

| FlexWrap : Layout.LayoutTypes.wrapType selector

| FlexDirection : Layout.LayoutTypes.flexDirection selector

| JustifyContent : Layout.LayoutTypes.justify selector

| AlignItems : Layout.LayoutTypes.align selector

| Position : Layout.LayoutTypes.positionType selector

| BackgroundColor : Revery_Core.Color.t selector

| Height : int selector

| Top : int selector

| Right : int selector

| Bottom : int selector

| Left : int selector

| MarginTop : int selector

| MarginLeft : int selector

| MarginRight : int selector

| MarginBottom : int selector

| Margin : int selector

| MarginVertical : int selector

| MarginHorizontal : int selector

| Margin2 : Style.xy selector

| Margin4 : Style.coords selector

| PaddingTop : int selector

| PaddingLeft : int selector

| PaddingRight : int selector

| PaddingBottom : int selector

| Padding : int selector

| PaddingVertical : int selector

| PaddingHorizontal : int selector

| Padding2 : Style.xy selector

| Padding4 : Style.coords selector

| Overflow : Layout.LayoutTypes.overflow selector

| BorderTop : Style.Border.t selector

| BorderLeft : Style.Border.t selector

| BorderRight : Style.Border.t selector

| BorderBottom : Style.Border.t selector

| Border : Style.Border.t selector

| BorderHorizontal : Style.Border.t selector

| BorderVertical : Style.Border.t selector

| Transform : Transform.t list selector

| Opacity : float selector

| BoxShadow : Style.BoxShadow.properties selector

| Cursor : Revery_Core.MouseCursors.t option selector

val select : selectionType. Style.allProps list -> 'selectionType selector -> 'selectionType -> 'selectionType

This function pulls out (selects) a style property based on the matching Variant that is passed in it uses a GADT to correctly type the return value of the function by matching the variant passed in