Module Revery_UI_Components__.Tree

type status =

| Open

| Closed

type 'a content = {

data : 'a;

id : int;

status : status;

}

type 'a tree =

| Empty

| Node of 'a content * 'a tree list

val defaultNodeStyles : [> `FlexDirection of Revery_UI.Layout.LayoutTypes.flexDirection | `MarginVertical of int ] list

val findNode : int -> 'a tree -> 'a content option

val default : indent:int -> Stdlib.String.t content -> Revery_UI.React.node Brisk_reconciler.element

val renderTree : ?⁠indent:int -> nodeRenderer:(indent:int -> 'a content -> 'b) -> emptyRenderer:(int -> 'b) option -> 'a tree -> 'b list

*

parameter ~indent

How much to indent the current (sub)tree. *

@param ~renderer

is a function which determines how each node is rendered *

@param t

The tree to convert to a tree of JSX elements.

val make : tree:'a tree -> nodeRenderer:(indent:int -> 'a content -> Revery_UI.React.node Revery_UI.React.element) -> ?⁠emptyRenderer:(int -> Revery_UI.React.node Revery_UI.React.element) option -> unit -> Revery_UI.viewNode Brisk_reconciler.element