Module Revery_UI_Hooks

module SpringHook = Revery_UI_Hooks__.Spring

include Revery_UI_Hooks__.Effect

val effect : 'a Revery_UI.React.Hooks.Effect.condition -> (unit -> (unit -> unit) option) -> ('a Revery_UI.React.Hooks.Effect.t -> 'b, 'c) Revery_UI.React.Hooks.t -> unit * ('b, 'c) Revery_UI.React.Hooks.t

Description:

The Effect Hook lets you perform side effects in function components.

Usage:

let%hook () = Hooks.effect(OnMount, () => {
  // Some side-effect

  // Clean-up function
  None;
});

include Revery_UI_Hooks__.Reducer

val reducer : initialState:'a -> ('b -> 'a -> 'a) -> ('a Revery_UI.React.Hooks.Reducer.t -> 'c, 'd) Revery_UI.React.Hooks.t -> ('a * ('b -> unit)) * ('c, 'd) Revery_UI.React.Hooks.t

Description:

An alternative to Hooks.state. Accepts a reducer of type (action, state) => newState, and returns the current state paired with a dispatch function.

Usage:

let%hook (state, dispatch) = Hooks.reducer(~initialState={counter:0}, reduce);

include Revery_UI_Hooks__.Ref

val ref : 'a -> ('a Stdlib.ref -> 'b, 'c) Revery_UI.React.Hooks.t -> 'a Stdlib.ref * ('b, 'c) Revery_UI.React.Hooks.t

include Revery_UI_Hooks__.State

val state : 'a -> ('a Revery_UI.React.Hooks.State.t -> 'b, 'c) Revery_UI.React.Hooks.t -> ('a * (('a -> 'a) -> unit)) * ('b, 'c) Revery_UI.React.Hooks.t

Description:

TODO

Usage:

let%hook (state, setState) = Hooks.state(0);

include Revery_UI_Hooks__.Tick

module Time = Revery_Core.Time

val tick : ?⁠tickRate:Time.t -> name:string -> (Time.t -> unit) -> ((Time.t -> unit) Stdlib.ref -> Revery_UI.React.Hooks.Effect.onMount Revery_UI.React.Hooks.Effect.t -> 'a, 'b) Revery_UI.React.Hooks.t -> unit * ('a, 'b) Revery_UI.React.Hooks.t

val time : ?⁠tickRate:Revery_UI_Hooks__.Tick.Time.t -> name:string -> unit -> (Revery_UI_Hooks__.Tick.Time.t Revery_UI.React.Hooks.Reducer.t -> (Revery_UI_Hooks__.Tick.Time.t -> unit) Stdlib.ref -> Revery_UI.React.Hooks.Effect.onMount Revery_UI.React.Hooks.Effect.t -> 'a, 'b) Revery_UI.React.Hooks.t -> Revery_UI_Hooks__.Tick.Time.t * ('a, 'b) Revery_UI.React.Hooks.t

val timer : ?⁠tickRate:Revery_UI_Hooks__.Tick.Time.t -> ?⁠active:bool -> name:string -> unit -> (Revery_UI_Hooks__.Tick.Time.t Revery_UI.React.Hooks.Reducer.t -> Revery_UI_Hooks__.Tick.Time.t Stdlib.ref -> (unit -> unit) option Stdlib.ref -> bool Revery_UI.React.Hooks.Effect.t -> 'a, 'b) Revery_UI.React.Hooks.t -> (Revery_UI_Hooks__.Tick.Time.t * (unit -> unit)) * ('a, 'b) Revery_UI.React.Hooks.t

val animation : ?⁠active:bool -> ?⁠onComplete:(unit -> unit) -> name:string -> 'a Revery_UI.Animation.t -> (bool Revery_UI.React.Hooks.State.t -> Revery_UI_Hooks__.Tick.Time.t Revery_UI.React.Hooks.Reducer.t -> Revery_UI_Hooks__.Tick.Time.t Stdlib.ref -> (unit -> unit) option Stdlib.ref -> bool Revery_UI.React.Hooks.Effect.t -> 'b, 'c) Revery_UI.React.Hooks.t -> ('a * Revery_UI.Animation.state * (unit -> unit)) * ('b, 'c) Revery_UI.React.Hooks.t

module Internal : sig ... end

val transition : ?⁠active:bool -> ?⁠duration:Time.t -> ?⁠delay:Time.t -> ?⁠easing:Revery_UI.Easing.t -> ?⁠initialValue:float -> float -> name:string -> ((float * float * Time.t) Revery_UI.React.Hooks.State.t -> bool Revery_UI.React.Hooks.State.t -> Revery_UI_Hooks__.Tick.Time.t Revery_UI.React.Hooks.Reducer.t -> Revery_UI_Hooks__.Tick.Time.t Stdlib.ref -> (unit -> unit) option Stdlib.ref -> bool Revery_UI.React.Hooks.Effect.t -> float Revery_UI.React.Hooks.Effect.t -> 'a, 'b) Revery_UI.React.Hooks.t -> float * ('a, 'b) Revery_UI.React.Hooks.t

val transitionf : ?⁠active:bool -> ?⁠delay:Time.t -> ?⁠easing:Revery_UI.Easing.t -> initialDuration:Time.t -> duration:(current:float -> target:float -> Time.t) -> ?⁠initialValue:float -> float -> name:string -> ((float * float * Time.t) Revery_UI.React.Hooks.State.t -> bool Revery_UI.React.Hooks.State.t -> Revery_UI_Hooks__.Tick.Time.t Revery_UI.React.Hooks.Reducer.t -> Revery_UI_Hooks__.Tick.Time.t Stdlib.ref -> (unit -> unit) option Stdlib.ref -> bool Revery_UI.React.Hooks.Effect.t -> float Revery_UI.React.Hooks.Effect.t -> 'a, 'b) Revery_UI.React.Hooks.t -> float * ('a, 'b) Revery_UI.React.Hooks.t

val spring : ?⁠enabled:bool -> name:string -> target:float -> ?⁠initialState:SpringHook.Spring.t -> ?⁠restThreshold:float -> SpringHook.Spring.Options.t -> (SpringHook.Spring.t Stdlib.ref -> Revery_UI_Hooks__.Tick.Time.t Revery_UI.React.Hooks.Reducer.t -> Revery_UI_Hooks__.Tick.Time.t Stdlib.ref -> (unit -> unit) option Stdlib.ref -> bool Revery_UI.React.Hooks.Effect.t -> 'a, 'b) Revery_UI.React.Hooks.t -> (float * (float -> unit)) * ('a, 'b) Revery_UI.React.Hooks.t

val mouseCapture : ?⁠onMouseDown:('a -> Revery_UI.NodeEvents.mouseButtonEventParams -> 'a option) -> ?⁠onMouseUp:('a -> Revery_UI.NodeEvents.mouseButtonEventParams -> 'a option) -> ?⁠onMouseMove:('a -> Revery_UI.NodeEvents.mouseMoveEventParams -> 'a option) -> ?⁠onMouseWheel:('a -> Revery_UI.NodeEvents.mouseWheelEventParams -> 'a option) -> ?⁠onRelease:('a option -> unit) -> unit -> ('a option Stdlib.ref -> Revery_UI.React.Hooks.Effect.onMount Revery_UI.React.Hooks.Effect.t -> 'b, 'c) Revery_UI.React.Hooks.t -> (('a -> unit) * 'a option) * ('b, 'c) Revery_UI.React.Hooks.t

* let%hook (captureMouse, captureState) = Hooks.mouseCapture(...); * * where captureMouse requires the initial state to be passed as its argument. * * Each callback is passed the current state and event data, and requires either a * state value to be returned or None to stop the capture.

val hover : unit -> (bool Revery_UI.React.Hooks.State.t -> 'a, 'b) Revery_UI.React.Hooks.t -> (bool * ('c -> unit) * ('d -> unit)) * ('a, 'b) Revery_UI.React.Hooks.t