Module Revery_UI__.Ui

type renderFunction = Revery_UI.React.reveryNode Revery_UI__.React.element -> unit

val start : Revery_Core.Window.t -> Revery_UI.React.reveryNode Revery_UI__.React.element -> renderFunction

start is the entry point for creating a UI driven by the brisk-reconciler React infrastructure.

This does the following:

Hook up a render loop

Take a React.element and render it via GPU

Hook up GLFW events and convert them to Revery UI events

start returns a function that can be used to update the UI.

An example usage would be: let update = start(window, <View />); update(<Button />);

val getActiveWindow : unit -> Revery_Core.Window.t option